+++ test/cljs/reitit/frontend/core_test.cljs
@@ -1,13 +1,16 @@
 (ns reitit.frontend.core-test
-  (:require [clojure.test :refer [deftest testing is are]]
+  (:require [clojure.string :as str]
+            [clojure.test :refer [are deftest is testing]]
+            [malli.core :as m]
+            [malli.transform :as mt]
+            [reitit.coercion :as rc]
+            [reitit.coercion.malli :as rcm]
+            [reitit.coercion.schema :as rcs]
             [reitit.core :as r]
             [reitit.frontend :as rf]
-            [reitit.coercion :as rc]
-            [schema.core :as s]
-            [reitit.coercion.schema :as rcs]
-            [reitit.coercion.malli :as rcm]
             [reitit.frontend.test-utils :refer [capture-console]]
-            [reitit.impl :as impl]))
+            [reitit.impl :as impl]
+            [schema.core :as s]))
 
 (deftest query-params-test
   (is (= {:foo "1"}
@@ -297,3 +300,33 @@
   (testing "Fragment encoding"
     (is (= "foo#foo+bar+%25"
            (rf/match->path {:path "foo"} nil "foo bar %")))))
+
+(deftest match->path-coercion-test
+  (testing "default keyword to string"
+    (is (str/starts-with?
+          (rf/match->path {:path "foo"} {:q :x})
+          "foo?q=x")))
+
+  (testing "default string transformer"
+    (is (= "foo?q=__x"
+           (rf/match->path {:data {:coercion rcm/coercion
+                                   :parameters {:query [[:map
+                                                         [:q {:decode/string (fn [s] (keyword (subs s 2)))
+                                                              :encode/string (fn [k] (str "__" (name k)))}
+                                                          :keyword]]]}}
+                            :path "foo"}
+                           {:q "x"}))))
+
+  (testing "custom string transformer"
+    (is (= "foo?q=--x"
+           (rf/match->path {:data {:coercion (rcm/create (assoc-in rcm/default-options
+                                                                   [:transformers :string :default]
+                                                                   (mt/transformer
+                                                                     {:name :foo-string
+                                                                      :encoders {:foo/type {:leave (fn [x] (str "--" x))}}})))
+                                   :parameters {:query [[:map
+                                                         [:q (m/-simple-schema
+                                                               {:type :foo/type
+                                                                :pred string?})]]]}}
+                            :path "foo"}
+                           {:q "x"})))))